+++ projects/ngx-panzoom/src/lib/panzoom-api.ts
@@ -22,4 +22,6 @@ export interface PanZoomAPI {
   centerBottomLeft: Function;
   centerTopRight: Function;
   centerBottomRight: Function;
+  updateContentDimensions: Function;
+  detectContentDimensions: Function;
 }